+++ genericsuite/util/jwt.py
@@ -1,7 +1,7 @@
 """
 JWT Library
 """
-from typing import Callable, Union
+from typing import Callable, Union, Optional
 import os
 import base64
 import datetime
@@ -39,7 +39,7 @@ PARAMS_FILE_USER_FILENAME_TEMPLATE = os.environ.get(
     'PARAMS_FILE_USER_FILENAME_TEMPLATE', 'params_[user_id].json')
 
 INVALID_TOKEN_ERROR_MESSAGE = 'Token is invalid'
-USER_ID_REQUIRED_ERROR_MESSAGE = 'User ID is required to get access token data'
+OTHER_TOKEN_ERROR_MESSAGE = 'Token error'
 
 
 class AuthTokenPayload(BaseModel):
@@ -123,7 +123,10 @@ def generate_access_token(length=64):
     return secrets.token_hex(length)
 
 
-def get_access_token_data(access_token: str, user_id: str = None) -> dict:
+def get_access_token_data(
+        access_token: str,
+        user_id: Optional[str] = None
+) -> dict:
     """
     Get the access token data from the database.
 
@@ -134,26 +137,30 @@ def get_access_token_data(access_token: str, user_id: str = None) -> dict:
     Returns:
         dict: The access token data.
     """
-    if not user_id:
-        return error_resultset(USER_ID_REQUIRED_ERROR_MESSAGE)
-    dbo = GenericDbHelperSuper(json_file="users")
-    user_data = dbo.fetch_row_raw(user_id)
-    if user_data['error']:
-        return error_resultset(user_data['error_message'])
     try:
-        # Check if the access token is valid
-        user_data['valid_token'] = False
-        _ = DEBUG and log_debug(
-            f'JWT.py | get_access_token_data | user_data: {user_data}')
-        resultset = user_data['resultset']
-        for api_key in resultset.get('users_api_keys', []):
-            if api_key['access_token'] == access_token and \
-               api_key['active'] == '1':
-                user_data['valid_token'] = True
-                break
-        if user_data['valid_token']:
-            user_data['user_id'] = get_id_as_string(resultset)
-            return user_data
+        dbo = GenericDbHelperSuper(json_file="users_api_keys")
+        api_key_data = dbo.fetch_row_raw(access_token)
+        if api_key_data['error']:
+            return error_resultset(api_key_data['error_message'])
+        resultset = api_key_data['resultset']
+        if not resultset:
+            return error_resultset(INVALID_TOKEN_ERROR_MESSAGE)
+        if user_id and resultset['user_id'] != user_id:
+            return error_resultset(INVALID_TOKEN_ERROR_MESSAGE)
+        if resultset['active'] != '1':
+            return error_resultset(INVALID_TOKEN_ERROR_MESSAGE)
+
+        user_id = resultset['user_id']
+        dbo = GenericDbHelperSuper(json_file="users")
+        user_data = dbo.fetch_row_raw(user_id)
+        if user_data['error']:
+            return error_resultset(user_data['error_message'])
+        if user_data['resultset'].get('active') != '1':
+            return error_resultset(INVALID_TOKEN_ERROR_MESSAGE)
+        user_data['valid_token'] = True
+        user_data['user_id'] = get_id_as_string(resultset)
+        return user_data
+
     except Exception as err:
         _ = DEBUG and log_error(
             'REQUEST_AUTHENTICATION@get_api_key_auth | Exception:'
@@ -163,23 +170,25 @@ def get_access_token_data(access_token: str, user_id: str = None) -> dict:
 
 def get_api_key_auth(
     request: Request,
-    user_id: str,
-    access_token: str
+    access_token: str,
+    user_id: Optional[str] = None,
 ) -> Union[AuthorizedRequest, dict]:
     """
     Get the authorized request from the request object.
 
     Args:
         request (Request): The request object.
+        access_token (str): The access token (with or without the
+            'Bearer ' prefix). Required for API Keys.
         user_id (str): The user ID specified as customer_id in the POST
-            body.
-        access_token (str): The access token (without 'Bearer ' prefix).
-            (required for API Keys)
+            body. Optional for API Keys.
 
     Returns:
         Union[AuthorizedRequest, dict]: The authorized request or an error
             message.
     """
+    if access_token.startswith('Bearer '):
+        access_token = access_token[7:]
     access_token_data = get_access_token_data(access_token, user_id)
     if access_token_data['error']:
         return access_token_data
@@ -233,6 +242,12 @@ def get_general_authorized_request(request: Request
             'REQUEST_AUTHENTICATION@get_general_authorized_request'
             f' | authorized_request = {authorized_request}')
     except Exception as err:
+        if "invalid token" not in str(err).lower():
+            log_error(
+                f'REQUEST_AUTHENTICATION@get_general_authorized_request'
+                f' | Exception: {str(err)}')
+            return standard_error_return(
+                OTHER_TOKEN_ERROR_MESSAGE)
         # API Keys processing
         project_id = request.headers.get('x-project-id', '')
         _ = DEBUG and log_debug(
@@ -241,12 +256,7 @@ def get_general_authorized_request(request: Request
             f'\n | jwt_token = {jwt_token}'
             f'\n | headers = {request.headers}'
             f'\n | request = {request}')
-        if project_id and jwt_token:
-            return get_api_key_auth(request, project_id, jwt_token)
-        _ = DEBUG and log_error(
-            f'REQUEST_AUTHENTICATION@get_general_authorized_request'
-            f' | Exception: {str(err)}')
-        return standard_error_return(INVALID_TOKEN_ERROR_MESSAGE)
+        return get_api_key_auth(request, jwt_token, project_id)
     return authorized_request